+++ gateway/ln-gateway/src/bin/cln_extension.rs
@@ -14,7 +14,7 @@ use cln_plugin::{options, Builder, Plugin};
 use cln_rpc::model;
 use cln_rpc::primitives::ShortChannelId;
 use fedimint_core::task::TaskGroup;
-use fedimint_core::Amount;
+use ln_gateway::cln::HtlcAccepted;
 use ln_gateway::gatewaylnrpc::complete_htlcs_request::{Action, Cancel, Settle};
 use ln_gateway::gatewaylnrpc::gateway_lightning_server::{
     GatewayLightning, GatewayLightningServer,
@@ -26,7 +26,6 @@ use ln_gateway::gatewaylnrpc::{
     SubscribeInterceptHtlcsResponse,
 };
 use secp256k1::PublicKey;
-use serde::{Deserialize, Deserializer, Serialize};
 use thiserror::Error;
 use tokio::io::{stdin, stdout};
 use tokio::sync::{mpsc, oneshot, Mutex};
@@ -70,44 +69,6 @@ async fn main() -> Result<(), anyhow::Error> {
         .map_err(|e| ClnExtensionError::Error(anyhow!("Failed to start server, {:?}", e)))?;
 
     Ok(())
-}
-
-/// The core-lightning `htlc_accepted` event's `amount` field has a "msat"
-/// suffix
-fn as_fedimint_amount<'de, D>(amount: D) -> Result<Amount, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let amount = String::deserialize(amount)?;
-    Ok(Amount::from_msats(
-        amount[0..amount.len() - 4].parse::<u64>().unwrap(),
-    ))
-}
-
-// TODO: upstream these structs to cln-plugin
-// See: https://github.com/ElementsProject/lightning/blob/master/doc/PLUGINS.md#htlc_accepted
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct Htlc {
-    #[serde(deserialize_with = "as_fedimint_amount")]
-    pub amount_msat: Amount,
-    // TODO: use these to validate we can actually redeem the HTLC in time
-    pub cltv_expiry: u32,
-    pub cltv_expiry_relative: u32,
-    pub payment_hash: bitcoin_hashes::sha256::Hash,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct Onion {
-    #[serde(default)]
-    pub short_channel_id: Option<String>,
-    #[serde(deserialize_with = "as_fedimint_amount")]
-    pub forward_msat: Amount,
-}
-
-#[derive(Clone, Serialize, Deserialize, Debug)]
-pub struct HtlcAccepted {
-    pub htlc: Htlc,
-    pub onion: Onion,
 }
 
 pub struct ClnRpcClient {}
